refactor: 共有設定パッケージを @trend-diary/config に一般化する
ツール別に vitest-config を切り出すと @trend-diary/* が乱立するため、
共有設定は単一の config パッケージへ集約し、サブパスで公開する方針に改める。
カバレッジ preset は @trend-diary/config/vitest として公開し、将来 tsconfig 等の
共有設定も同パッケージに追加できるようにする。

- packages/vitest-config を packages/config へリネームし、
  src/index.ts を src/vitest.ts へ移す
- exports を "./vitest" サブパスに変更する
- 利用側7ファイルの import と devDependencies を @trend-diary/config へ更新する
- 構成ツリーの記載を更新する
