+++ packages/config/tsconfig.json
