+++ packages/datastore/package.json
@@ -22,7 +22,7 @@
   },
   "devDependencies": {
     "@cloudflare/workers-types": "catalog:",
-    "@trend-diary/vitest-config": "workspace:*",
+    "@trend-diary/config": "workspace:*",
     "@types/node": "catalog:",
     "drizzle-kit": "0.31.10",
     "vitest": "catalog:"
