+++ packages/logger/package.json
@@ -14,7 +14,7 @@
     "pino": "catalog:"
   },
   "devDependencies": {
-    "@trend-diary/vitest-config": "workspace:*",
+    "@trend-diary/config": "workspace:*",
     "@types/node": "catalog:",
     "vitest": "catalog:"
   }
